chore: blogposts broken into own component
