+++ src/components/Blogpost/blogpost.module.scss
@@ -1,0 +1,29 @@
+.blog-post {
+  h3 {
+    margin-bottom: 0.25rem;
+  }
+  p {
+    margin-bottom: 0.5rem;
+    display: -webkit-box;
+    -webkit-line-clamp: 3;
+    -webkit-box-orient: vertical;
+    overflow: hidden;
+  }
+
+  span {
+    font-size: 0.875rem;
+    display: inline-block;
+    margin-right: 0.75rem;
+    margin-bottom: 0.5rem;
+  }
+
+  a {
+    text-decoration: none;
+    .external-link {
+      margin-bottom: 4px;
+      vertical-align: middle;
+      margin-left: 12px;
+      width: 14px;
+    }
+  }
+}
