+++ src/components/Blogposts/blogposts.module.scss
@@ -7,37 +7,6 @@
   img {
     margin-bottom: 0;
   }
-
-  a {
-    text-decoration: none;
-    .external-link {
-      margin-bottom: 4px;
-      vertical-align: middle;
-      margin-left: 12px;
-      width: 14px;
-    }
-  }
-}
-
-.blog-post {
-  margin-bottom: 2rem;
-
-  h3 {
-    margin-bottom: 0.25rem;
-  }
-  p {
-    margin-bottom: 0.5rem;
-    display: -webkit-box;
-    -webkit-line-clamp: 3;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-  }
-  span {
-    font-size: 0.875rem;
-    display: inline-block;
-    margin-right: 0.75rem;
-    margin-bottom: 0.5rem;
-  }
 }
 
 .flex {
